Removed Test project and moved all tests to its own project
