+++ BankManager_Csharp/BankManager_Csharp.sln
@@ -5,9 +5,7 @@ VisualStudioVersion = 15.0.26430.14
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BankManager_Csharp", "BankManager_Csharp\BankManager_Csharp.csproj", "{179E123D-D807-4800-908C-FD17D7E2F693}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BankManager_Csharp_Test", "BankManager_Csharp_Test\BankManager_Csharp_Test.csproj", "{21A6D5D0-972D-411D-AF79-51A5822FDF39}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BankManager_Csharp_Console", "BankManager_Csharp_Console\BankManager_Csharp_Console.csproj", "{C21CA6CF-131A-46F7-8678-53EE66943971}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BankManager_Csharp_Console", "BankManager_Csharp_Console\BankManager_Csharp_Console.csproj", "{C21CA6CF-131A-46F7-8678-53EE66943971}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,10 +17,6 @@ Global
 		{179E123D-D807-4800-908C-FD17D7E2F693}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{179E123D-D807-4800-908C-FD17D7E2F693}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{179E123D-D807-4800-908C-FD17D7E2F693}.Release|Any CPU.Build.0 = Release|Any CPU
-		{21A6D5D0-972D-411D-AF79-51A5822FDF39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{21A6D5D0-972D-411D-AF79-51A5822FDF39}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{21A6D5D0-972D-411D-AF79-51A5822FDF39}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{21A6D5D0-972D-411D-AF79-51A5822FDF39}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C21CA6CF-131A-46F7-8678-53EE66943971}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C21CA6CF-131A-46F7-8678-53EE66943971}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C21CA6CF-131A-46F7-8678-53EE66943971}.Release|Any CPU.ActiveCfg = Release|Any CPU
