+++ BankManager_Csharp/BankManager_Csharp/packages.config
@@ -8,4 +8,5 @@
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net452" />
   <package id="Modernizr" version="2.6.2" targetFramework="net452" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net452" />
+  <package id="NUnit" version="3.7.1" targetFramework="net452" />
 </packages>
