--- BankManager_Csharp/BankManager_Csharp_Test/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net452" />
-  <package id="NUnit" version="3.7.1" targetFramework="net452" />
-</packages>
